requirements.txt
fastapi
uvicorn
openai-whisper
transformers
torch
python-multipart
tf-keras